Append with CRIDs added to portal since 6/5/16
I've appended the CSV with the 5 CRIDs added to portal.ipra.org between 6/5 and 6/25/16:
1074270
1074294
1080064
1080144
1080972

GitHub is telling me, "We've detected the file encoding as ISO-8859-1. When you commit changes we will transcode it to UTF-8." That's why it's showing as 108 items added when it's really only the 5 lines at the end.