Append with files for CRIDs added to portal
I've appended the CSV with 10 files related to 4 of the 5 CRIDs added to portal.ipra.org between 6/5 and 6/25/16 (1 CRID had zero files):
1074270
1074294
1080144
1080972

GitHub is telling me, "We've detected the file encoding as ISO-8859-1. When you commit changes we will transcode it to UTF-8." That's why it's showing as many items added when it's really only the 10 lines at the end.

I *did not* review the existing CRIDs to see if new files were added for them.